requirements.txt
altgraph==0.15
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Faker==0.8.13
future==0.16.0
idna==2.6
kafka-python==1.4.2
macholib==1.9
numpy==1.14.2
oauthlib==2.0.7
pkg-resources==0.0.0
psycopg2==2.7.4
python-dateutil==2.7.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
text-unidecode==1.2
url-normalize==1.3.3
urllib3==1.22
